sessions/ChatBot/requirements.txt
python-dotenv
groq
langchain-groq
langchain-core
streamlit
